Add default parameters to data store interfaces (#30)
Prior to this commit the functions within `FeatureFlagDataStore` and
`ObservableFeatureFlagDataStore` did not take default parameters.
Instead, this was handled one level up in the `FeatureFlagManager` and
`ObservableFeatureFlagManager` implementations. This option was good for
simplicity, but presented one problem.

The LaunchDarkly SDK requires that customers of their API set certain
configurations in order to reliably use their client APIs which return
an `EvaluationDetail<T>`. Prior to this commit we used the evaluation of
flags to determine if the static default value we passed into the LD
client API was returned to us. This gave us to signal we needed to then
return the `FeatureFlag#default` value that this library guarantees API
consumers will have returned when no value is available.

It's unreasonable for us to expect all of our API consumers who use LD
would properly configure their client for our implementation needs.
Rather than enforcing that requirement, this commit alters the API
surface of our data stores such that they have knowledge of what the
`FeatureFlag#default` value is. Instead of having the feature flag
managers responsible for returning defaults directly from flags, they
now query defaults and pass them along to a data store. This gives
implementations the flexibility to pass the default value along to the
underlying SDK in use, or use some other method for determining if a
default value needs to be returned.

References #4